requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
feedparser==6.0.10
google-generativeai==0.8.6
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9